Move classical to be a variant.
Classical is now a variant that most other variants will be based on.
