+++ variants/classical/bindata.go
