+++ variants/classical/common/common.go
