+++ variants/classical/datc/datc_v2.4_06.txt
