+++ variants/classical/datc/dipai.txt
