+++ variants/classical/datc/diplicity_errors.txt
