+++ variants/classical/datc/droidippy_errors.txt
