+++ variants/classical/droidippy/extract_droidippy_game.rb
