+++ variants/classical/start/supply_centers.go
