+++ variants/franceaustria/franceaustria.go
@@ -1,13 +1,13 @@
 package franceaustria
 
 import (
-	"github.com/zond/godip/classical"
-	"github.com/zond/godip/classical/orders"
-	"github.com/zond/godip/classical/start"
 	"github.com/zond/godip/state"
+	"github.com/zond/godip/variants/classical"
+	"github.com/zond/godip/variants/classical/orders"
+	"github.com/zond/godip/variants/classical/start"
 	"github.com/zond/godip/variants/common"
 
-	cla "github.com/zond/godip/classical/common"
+	cla "github.com/zond/godip/variants/classical/common"
 	dip "github.com/zond/godip/common"
 )
